requirements.txt
# FastAPI
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
email-validator>=2.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.11.0
pymysql>=1.0.3
cryptography>=40.0.0

# Authentication
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1

# Task Queue
celery>=5.3.0
redis>=4.5.0

# Testing
pytest>=7.3.0
httpx>=0.24.0
pytest-asyncio>=0.21.0

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.0
aiofiles>=23.1.0